+++ application/controllers/Disciplina.php
@@ -1,37 +1,43 @@
 <?php
-	class Disciplina extends CI_Controller {
+	require_once("Geral.php");
+	
+	class Disciplina extends Geral {
 		/*
 			no construtor carregamos as bibliotecas necessarias e tambem nossa model
 		*/
 		public function __construct()
 		{
 			parent::__construct();
-			
-			$this->load->model('login_model');
-			$this->load->model('Disciplina_model');
-			$this->load->model('Categoria_model');
-			$this->load->helper('url_helper');
-			$this->load->helper('url');
-			$this->load->helper('html');
-			$this->load->helper('form');
-			$this->load->library('session');
-			//verifica se o usuario este logado, a sessao alem de existir tambem deve ser valida
 			if(empty($this->login_model->session_is_valid($this->session->id)['id']))
 				redirect('login/login');
+			$this->load->model('Disciplina_model');
+			$this->load->model('Categoria_model');
+			$this->set_menu();
+			$this->data['controller'] = get_class($this);
+			$this->data['menu_selectd'] = $this->Geral_model->get_identificador_menu(strtolower(get_class($this)));
 		}
+		
 		
 		/*
 			listar as disciplinas cadastradas
 		*/
-		public function index(){
-			$data['url'] = base_url();
-			$data['controller'] = 'disciplina';
-			$data['Disciplinas'] = $this->Disciplina_model->get_disciplina();
-			$data['title'] = 'Administração';
-			$data['message'] = 'Administração';
-			$this->load->view('templates/header_admin',$data);
-			$this->load->view('disciplina/index',$data);
-			$this->load->view('templates/footer',$data);
+		public function index($page = false)
+		{
+			if($page === false)
+				$page = 1;
+			
+			$this->data['title'] = 'Administração';
+			if($this->Geral_model->get_permissao(READ,get_class($this)) == true)
+			{
+				$this->data['Disciplinas'] = $this->Disciplina_model->get_disciplina(false,$page);
+				
+				$this->data['paginacao']['size'] = $this->data['Disciplinas'][0]['size'];
+				$this->data['paginacao']['pg_atual'] = $page;
+				
+				$this->view("disciplina/index",$this->data);
+			}
+			else
+				$this->view("templates/permissao",$this->data);
 		}
 		
 		/*
@@ -40,38 +46,53 @@
 		*/
 		public function deletar($id = null)
 		{
-			$this->Disciplina_model->delete_disciplina($id);
+			if($this->Geral_model->get_permissao(DELETE,get_class($this)) == true)
+				$this->Disciplina_model->delete_disciplina($id);
+			else
+				$this->view("templates/permissao",$this->data);
 		}
 		
 		/*
 			APAGAR UMA DISCIPLINA DESDE QUE EXISTA A SESSAO DE USUARIO E A MESMA
 			SEJA VALIDA
 		*/
-		public function create_edit($id = null)
+		public function edit($id = null)
 		{
-			$data['url'] = base_url();
+			$this->data['title'] = 'Administração';
+			if($this->Geral_model->get_permissao(UPDATE,get_class($this)) == true)
+			{
+				$this->data['Disciplina'] = $this->Disciplina_model->get_disciplina($id);
+				$this->data['Categoria'] = $this->Categoria_model->get_categoria();
+				$this->view("disciplina/create_edit",$this->data);
+			}
+			else
+				$this->view("templates/permissao",$this->data);
+		}
 
-			$data['Disciplina'] = $this->Disciplina_model->get_disciplina($id);
-			
-			$data['Categoria'] = $this->Categoria_model->get_categoria();
-			$data['title'] = 'Administração';
-			$data['controller'] = 'disciplina';
-			$data['message'] = 'Administração';
-			$this->load->view('templates/header_admin',$data);
-			$this->load->view('disciplina/create_edit',$data);
-			$this->load->view('templates/footer',$data);
+		public function create($id = null)
+		{
+			$this->data['title'] = 'Administração';
+			if($this->Geral_model->get_permissao(CREATE,get_class($this)) == true)
+			{
+
+				$this->data['Disciplina'] = $this->Disciplina_model->get_disciplina($id);
+				$this->data['Categoria'] = $this->Categoria_model->get_categoria();
+				$this->view("disciplina/create_edit",$this->data);
+			}
+			else
+				$this->view("templates/permissao",$this->data);
 		}
 		
 		public function store()
 		{
 			$dataToSave = array(
-				'Id' => $this->input->post('Id'),
-				'Ativo' => 1,
-				'Nome' => $this->input->post('Nome'),
-				'CategoriaId' => $this->input->post('CategoriaId')
+				'id' => $this->input->post('id'),
+				'ativo' => 1,
+				'nome' => $this->input->post('nome'),
+				'categoria_id' => $this->input->post('categoria_id')
 			);
 			//bloquear acesso direto ao metodo store
-			 if(!empty($dataToSave['Nome']))
+			 if(!empty($dataToSave['nome']))
 				$this->Disciplina_model->set_disciplina($dataToSave);
 			 else
 				redirect('admin/dashboard');
